Fix dependency ordering with synthetic timer reactions
Previously we had one reaction for both the
bootstrap and the reschedule. However, the
bootstrap may schedule reactions in the current
tag, so it should be upstream of those reactions.
But since this reaction is synthesized with an
index greater than any user-defined reaction,
there was a priority edge that created a cycle.

The solution now, is to split the timer reaction
into a bootstrap reaction and a reschedule reaction,
and to abolish priority edges between synthetic reactions

Compare the current, fixed diagram:

digraph {
    0 [ label = "Special(Startup)" ]
    1 [ label = "Special(Shutdown)" ]
    2 [ label = "Timer(/t)" ]
    3 [ label = "Timer(/t2)" ]
    4 [ label = "Reaction(/0)" ]
    5 [ label = "Reaction(/1)" ]
    6 [ label = "Reaction(/2)" ]
    7 [ label = "Reaction(/3@reschedule_t)" ]
    8 [ label = "Reaction(/4@reschedule_t2)" ]
    9 [ label = "Reaction(/5@bootstrap_t)" ]
    10 [ label = "Reaction(/6@bootstrap_t2)" ]
    4 -> 5 [ ]
    5 -> 6 [ ]
    3 -> 4 [ ]
    2 -> 5 [ ]
    1 -> 6 [ ]
    2 -> 7 [ ]
    0 -> 9 [ ]
    9 -> 2 [ ]
    3 -> 8 [ ]
    0 -> 10 [ ]
    10 -> 3 [ ]
}

to the old one:

digraph {
    0 [ label = "Special(Startup)" ]
    1 [ label = "Special(Shutdown)" ]
    2 [ label = "Timer(/t)" ]
    3 [ label = "Timer(/t2)" ]
    4 [ label = "Reaction(/0)" ]
    5 [ label = "Reaction(/1)" ]
    6 [ label = "Reaction(/2)" ]
    7 [ label = "Reaction(/3@schedule_t)" ]
    8 [ label = "Reaction(/4@schedule_t2)" ]
    4 -> 5 [ ]
    5 -> 6 [ ]
    6 -> 7 [ ]
    7 -> 8 [ ]
    3 -> 4 [ ]
    2 -> 5 [ ]
    1 -> 6 [ ]
    2 -> 7 [ ]
    0 -> 7 [ ]
    3 -> 8 [ ]
    0 -> 8 [ ]
}

(in that one, the edge between /3, /4 and their
respective downstream are missing, which was throwing
off the layerng algorithm and subsequent scheduling.

This example is the test Timers.lf