Change Readable/WritablePort to have same lifetime as Port
Code using writable/readable ports only deals in references
to them. This is more ergonomic, as it removes a lot of explicit borrows
needed to avoid moving the reference inadvertently. We can
also make reactions use an immutable vars for all ports.

Eg previously, we injected
  mut out: WritablePort<'_, T>

With that you can reassign out, even if the mut should only
be used to get an &mut. If you want not to move the port you
have to explicitly borrow with &mut.

Now the same will be injected as
  out: &mut WritablePort<T>

This behaves in a more expected manner (you have a reference
to the port and it behaves like a reference, not an owned object).

This also removes one level of indirection, as previously
an &mut WritablePort was actually a double reference to a Port.
Now it's practically a single reference, as the WritablePort *is*
the Port under a different name.